refactor: move the project detail components into components/project
