+++ frontend/src/components/project/index.ts
@@ -1,0 +1,6 @@
+export { OpinionForm, type OpinionFormProps } from "./OpinionForm";
+export { OtherOpinionsTable, type OtherOpinionsTableProps } from "./OtherOpinionsTable";
+export { ResultsSection, type ResultsSectionProps } from "./ResultsSection";
+export { TriangleVisualization, type TriangleVisualizationProps } from "./TriangleVisualization";
+export { TeamTable, type TeamTableProps } from "./TeamTable";
+export { MemberProfileDialog, type MemberProfileDialogProps } from "./MemberProfileDialog";
